Remove StepwisePipeline class
